Struct ion_rs::binary::non_blocking::raw_binary_reader::RawBinaryReader

pub struct RawBinaryReader<A: AsRef<[u8]> + Expandable> { /* private fields */ }

A raw binary reader that pulls input bytes from a fixed buffer.

If any read operation fails due to the buffer containing incomplete data, that method will return IonError::Incomplete.

If the buffer (generic type A) is a Vec<u8>, then data can be appended to it between read operations. This can be useful when reading from a data source that is growing over time, such as when tailing a growing file, reading over the network, or waiting for user input. Applications can read from the buffer until they encounter an Incomplete. Then, when more data is available, they can use read_from or append_bytes to add that data to the buffer. Finally, they can retry the read operation that had previously failed.

Note that if the buffer runs out of data between top level values, this will be interpreted as the end of the stream. Applications can still add more data to the buffer and resume reading.

Implementations

impl<A: AsRef<[u8]> + Expandable> RawBinaryReader<A>

pub fn new(source: A) -> RawBinaryReader<A>

Constructs a RawBinaryReader from a value that can be viewed as a byte slice.

pub fn annotations_iter( &self ) -> impl Iterator<Item = IonResult<RawSymbolToken>> + '_

Creates an iterator that lazily reads the VarUInt symbol IDs in this value’s annotations wrapper. If the reader is not on a value or the current value does not have annotations, the iterator will be empty.

pub fn read_symbol_id(&mut self) -> IonResult<SymbolId>

If the reader is currently positioned on a symbol value, parses that value into a SymbolId.

pub fn read_str_bytes(&mut self) -> IonResult<&[u8]>

If the reader is currently positioned on a string, returns the slice of bytes that represents that string’s UNVALIDATED utf-8 bytes. This method is available for performance optimization in scenarios where utf-8 validation may be unnecessary and/or a bottleneck. It is strongly recommended that you use read_str unless absolutely necessary.

pub fn read_blob_bytes(&mut self) -> IonResult<&[u8]>

If the reader is currently positioned on a blob, returns a slice containing its bytes.

pub fn read_clob_bytes(&mut self) -> IonResult<&[u8]>

If the reader is currently positioned on a clob, returns a slice containing its bytes.

pub fn header_length(&self) -> usize

pub fn raw_bytes(&self) -> Option<&[u8]>

Returns a slice containing the current value’s bytes. In the case of a container the raw bytes will consist of its field ID (if present), its annotations (if present), and its header. In the case of a non-container value, the bytes for the value itself is also included.

pub fn raw_field_id_bytes(&self) -> Option<&[u8]>

pub fn raw_header_bytes(&self) -> Option<&[u8]>

pub fn raw_value_bytes(&self) -> Option<&[u8]>

pub fn raw_annotations_bytes(&self) -> Option<&[u8]>

pub fn field_id_length(&self) -> Option<usize>

pub fn field_id_offset(&self) -> Option<usize>

pub fn field_id_range(&self) -> Option<Range<usize>>

pub fn annotations_length(&self) -> Option<usize>

pub fn annotations_offset(&self) -> Option<usize>

pub fn annotations_range(&self) -> Option<Range<usize>>

pub fn header_offset(&self) -> usize

pub fn header_range(&self) -> Range<usize>

pub fn value_length(&self) -> usize

pub fn value_offset(&self) -> usize

pub fn value_range(&self) -> Range<usize>