Struct ion_rs::binary::raw_binary_writer::RawBinaryWriter

pub struct RawBinaryWriter<W: Write> { /* private fields */ }

A system-level streaming binary Ion writer. This writer does not provide symbol table management; symbol-related operations (e.g. setting field IDs and annotations or writing symbol values) require a valid symbol ID to be provided by the caller.

To produce a valid binary Ion stream, the writer MUST call RawBinaryWriter::write_ion_version_marker before writing any data.

Implementations

impl<W: Write> RawBinaryWriter<W>

pub fn clear_annotations(&mut self)

pub fn has_annotations(&self) -> bool

pub fn write_symbol_id(&mut self, symbol_id: SymbolId) -> IonResult<()>

pub fn set_field_id(&mut self, field_id: SymbolId)

pub fn output(&self) -> &W

Returns a reference to the underlying io::Write implementation.

pub fn output_mut(&mut self) -> &mut W

Returns a mutable reference to the underlying io::Write implementation. Modifying the underlying sink is an inherently risky operation and can result in unexpected behavior. It is not recommended for most use cases.

pub fn add_annotation<A: AsRawSymbolTokenRef>(&mut self, annotation: A)