Trait ion_rs::data_source::ToIonDataSource

pub trait ToIonDataSource {
    type DataSource: IonDataSource;

    // Required method
    fn to_ion_data_source(self) -> Self::DataSource;
}

Types that implement this trait can be converted into an implementation of io::BufRead, allowing users to build a Reader from a variety of types that might not define I/O operations on their own.

Required Associated Types

type DataSource: IonDataSource

Required Methods

fn to_ion_data_source(self) -> Self::DataSource