Module ion_rs::element

Provides a dynamically typed, materialized representation of an Ion value.

An Element represents an (annotations, value) pair, where a value is an Ion integer, float, list, struct, etc.

For reference here are a couple other value style APIs for JSON:

simd_json’s Value API

serde_json’s Value API

Re-exports

pub use crate::types::Blob;

pub use crate::types::Bytes;

pub use crate::types::Clob;

pub use crate::types::List;

pub use crate::types::SExp;

pub use crate::types::Sequence;

pub use crate::types::Struct;

Modules

builders

element_stream_reader

element_stream_writer

reader

Provides APIs to read Ion data into Element from different sources such as slices or files.

writer

Provides utility to serialize Ion data from Element into common targets such as byte buffers or files.

Structs

Annotations

An ordered sequence of symbols that convey additional, application-specific information about their associated Ion value.

Element

An (annotations, value) pair representing an Ion value.

Enums

Value

Variants for all values within an Element.

Traits

IntoAnnotatedElement

Allows types that can be converted into an Ion Value to also specify annotations, producing an Element.

IntoAnnotations

Defines conversion into Annotations.