Trait ion_rs::element::writer::ElementWriter

pub trait ElementWriter {
    // Required method
    fn write_element(&mut self, element: &Element) -> IonResult<()>;

    // Provided method
    fn write_elements<'a, I: IntoIterator<Item = &'a Element>>(
        &'a mut self,
        elements: I
    ) -> IonResult<()> { ... }
}

Serializes Element instances into some kind of output sink.

Required Methods

fn write_element(&mut self, element: &Element) -> IonResult<()>

Serializes a single Element as a top-level value.

Provided Methods

fn write_elements<'a, I: IntoIterator<Item = &'a Element>>( &'a mut self, elements: I ) -> IonResult<()>

Serializes a collection of Element as a series of top-level values.

This will return Err if writing any element causes a failure.