Module ion_rs::result

Structs

Position

Position represents the location within an Ion stream where an error has been identified. For all formats byte_offset will contain the number of bytes into the stream that have been processed prior to encountering the error. When working with the text format, line_column will be updated to contain the line and column as well.

Enums

IonError

Represents the different types of high-level failures that might occur when reading Ion data.

Functions

decoding_error

A convenience method for creating an IonResult containing an IonError::DecodingError with the provided description text.

decoding_error_raw

A convenience method for creating an IonError::DecodingError with the provided operation text. Useful for calling Option#ok_or_else.

encoding_error

A convenience method for creating an IonResult containing an IonError::EncodingError with the provided description text.

encoding_error_raw

A convenience method for creating an IonError::EncodingError with the provided operation text. Useful for calling Option#ok_or_else.

illegal_operation

A convenience method for creating an IonResult containing an IonError::IllegalOperation with the provided operation text.

illegal_operation_raw

A convenience method for creating an IonError::IllegalOperation with the provided operation text. Useful for calling Option#ok_or_else.

incomplete_data_error

incomplete_data_error_raw

incomplete_text_error

incomplete_text_error_raw

Type Aliases

IonResult

A unified Result type representing the outcome of method calls that may fail.