Struct ion_rs::SystemReader

pub struct SystemReader<R: RawReader> { /* private fields */ }

A streaming Ion reader that:

maintains a symbol table

processes local symbol tables as they are encountered

can resolve symbol IDs into their associated text

does NOT filter out encoding data like IVMs and symbol tables

SystemReader itself is format-agnostic; all format-specific logic is handled by the wrapped RawReader implementation.

If the user skips over some or all of a local symbol table in the stream (via either next() or step_out()), the SystemReader will automatically process the rest of the local symbol table before advancing. This allows users to visit Ion values used to encode local symbol tables without also being responsible for interpreting them (as they would with an implementation of the RawReader trait.)

Implementations

impl<R: RawReader> SystemReader<R>

pub fn new(raw_reader: R) -> SystemReader<R>

pub fn raw_annotations(&self) -> impl Iterator<Item = RawSymbolToken> + '_

pub fn symbol_table(&self) -> &SymbolTable

pub fn read_raw_symbol(&mut self) -> IonResult<RawSymbolToken>

pub fn raw_field_name_token(&mut self) -> IonResult<RawSymbolToken>

pub fn current_string_was_consumed(&self) -> bool

impl<T: AsRef<[u8]> + Expandable> SystemReader<RawBinaryReader<T>>

Functionality that is only available if the data source we’re reading from is in-memory, like a Vec<u8> or &[u8].

pub fn raw_bytes(&self) -> Option<&[u8]>

pub fn raw_field_id_bytes(&self) -> Option<&[u8]>

pub fn raw_header_bytes(&self) -> Option<&[u8]>

pub fn raw_value_bytes(&self) -> Option<&[u8]>

pub fn raw_annotations_bytes(&self) -> Option<&[u8]>

pub fn field_id_length(&self) -> Option<usize>

pub fn field_id_offset(&self) -> Option<usize>

pub fn field_id_range(&self) -> Option<Range<usize>>

pub fn annotations_length(&self) -> Option<usize>

pub fn annotations_offset(&self) -> Option<usize>

pub fn annotations_range(&self) -> Option<Range<usize>>

pub fn header_length(&self) -> usize

pub fn header_offset(&self) -> usize

pub fn header_range(&self) -> Range<usize>

pub fn value_length(&self) -> usize

pub fn value_offset(&self) -> usize

pub fn value_range(&self) -> Range<usize>