Struct ion_rs::types::Symbol

pub struct Symbol { /* private fields */ }

The text of a fully resolved field name, annotation, or symbol value. If the symbol has known text (that is: the symbol is not $0), it will be stored as either a String or a shared reference to text in a symbol table.

Implementations

impl Symbol

pub fn owned<I: Into<String>>(text: I) -> Symbol

pub fn shared(text: Arc<str>) -> Symbol

pub fn unknown_text() -> Symbol

pub fn text(&self) -> Option<&str>

pub fn text_or_error(&self) -> IonResult<&str>